bot/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-content-organizer-bot"
version = "0.1.0"
description = "Telegram bot for content organization"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot[all]>=21.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=7.0.0",
    "black>=26.1.0",
    "ruff>=0.15.0",
]

[tool.black]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
